+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 
 ### Fixed
 - `wp` and `drush` commands both now use the object buffer and output the result of the operation without formatting. (#1023)
+- `Runner::getLogger()` will no longer return `null`. (#1049)
 
 ## [0.11.1] - 2016-03-30